+++ pyas2/__init__.py
@@ -1,2 +1,2 @@
 __title__ = 'pyAS2'
-__version__ = '0.3.8'
+__version__ = '0.4.0'